feat(spore): sovereign mode, SSH relay expansion, tab-inoculate UI
- sovereign.rs: Spore sovereign mode for independent operation
- ssh_relay: expanded relay with connection management
- eval_api: updated evaluation endpoints
- wasm_bindings: additional JS bridge methods
- tab-inoculate.js: new browser-side inoculation UI

+++ symthaea/crates/symthaea-spore/src/engine.rs
@@ -830,7 +830,7 @@ impl SporeEngine {
             timestamp_secs: std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH).map(|d| d.as_secs()).unwrap_or(0),
             consciousness_level, harmony_alignment,
             metacog_accuracy: 0.0, allostatic_load: 0.0,
-            dream_wisdom_count: self.dream_journal.count(),
+            dream_wisdom_count: self.dream_journal.count() as usize,
             coherence_score: consciousness_level,
             safety_level: self.immune.safety_level() as u8,
         });
